地址管理

+++ SXK/SXK/Sections/Promulgate/Controller/PromulgateVC.m
@@ -105,7 +105,7 @@
         [cell fillWithTitle:self.firstArr[indexPath.row]];
         return cell;
     }
-    NSString *CellIdentifier = [NSString stringWithFormat:@"cell%ld%ld",indexPath.section,indexPath.row];//不使用复用
+    NSString *CellIdentifier = [NSString stringWithFormat:@"cell%ld%ld",(long)indexPath.section,(long)indexPath.row];//不使用复用
     
     TypeCell *cell = [tableView dequeueReusableCellWithIdentifier:CellIdentifier];
     if (!cell) {
